[0007] cmangos-classic update z2051
Apply CREATURE_FLAG_EXTRA_NO_TALKTO_CREDIT to all creatures by
Schmoozerd
The flag was introduced in [s1442]. It is assumed that by now all
quests that require some talk-to credit are handled by different means.
If you catch quests that don't work yet, remove this flag for this npc
and report to your DB provider.

https://github.com/cmangos/mangos-classic/commit/c2f3d55f1803ffa8eaad29f
53f8f4c0c8723ebfb