[0022] Fixed DB errors
This fixes the numerous DB errors shown in log file when starting the
server. Most of the errors came from orphan content (probably from a
missed TBC/UDB backport), when relevant, the missing content was added
or fixed. Please, see the SQL update file for the complete list of what
was fixed (and how).

Note: the errors related to the condition table are not fixed by this
commit. They will be handle later. However, those errors have no impact
on server.