[0326] Released Classic DB version 1.2.1 Classic DB release version 1.2.1 'Blackhand Castle'
This release is compatible with revision z2401 of CMAnGOS-classic and comes with revision z2683 of SD2-classic and ACID-classic 1.0.4

Thanks to all the contributors for their reports, fixes and tests, especially amongst them: scotty0100, Neotmiren, Ela, Evildead666 and Metalica
Very special thanks to Schmoozerd and the UDB team as this release would not have been possible without their great and heavy contributions.

The main topic of this update is a total rework of Blackrock Spire with hundreds of fixes: creatures spawns (mostly), GO spawns, patrols, faction, models, flags and much more... The instance is now far more closer to was it was on retail.

Note that this release is a small intermediate version released with the closure of the milestone dedicated to fixing Blackrock Spire. The full change log will ship later when version 1.3.0 is released. In the meantime, all changes can be seen in the various commits, as always.