[0431] Released Classic DB version 1.3
Classic DB release version 1.3 'The Dread Citadel'

This release is compatible with revision z2436 of CMAnGOS-classic and
comes with revision z2689 of SD2-classic and with ACID-classic 1.0.5

Thanks to all the contributors for their reports, fixes and tests,
especially amongst them: scotty0100, Neotmiren, Ela, Swater96, wowpsp,
Shubu, Patman64 and Metalica
Very special thanks to the X-Savior, Xfurry, the TBC-DB and the UDB team
as this release would not have been possible without their great and
heavy contributions.

The main topic of this version is a big rework of Naxxramas raid
instance: with hundreds of fixes: creatures spawns, GO spawns, patrols,
spells, faction, models, flags and much more...
The dungeon is now far more closer to was it was on retail.

Many gossips, quests, items, NPCs and dungeons were fixed as usual.
Please, read the CHANGELOG.txt file for a full list of changes in
version 1.3 since version 1.2.