[0616] Released Classic DB version 1.4
Classic DB release version 1.4 'The Lost Kingdom'

This release is compatible with revision z2479 of CMaNGOS-classic and comes with revision z2695 of SD2-classic and with ACID-classic 1.0.5

Thanks to all the contributors for their reports, fixes and tests, especially amongst them: Tobschinski, evil@wow, scotty0100, Neotmiren, Sinoxan, Patman64, Metalica, Efonius, Krullgor, time-out...
As always, very special thanks to the ScriptDev2, ACID, TBC-DB and the UDB teams as this release would not have been possible without their great and heavy contributions. Classic DB is a cross databases project and many of the updates are possible thanks to the great work of our fellow developers from other projects.

The main topic of this version is a big rework of both Temple of Ahn'Qiraj and Ruins of Ahn'Qiraj raid instances, with hundreds of fixes: creatures spawns, GO spawns, patrols, spells, factions, models, flags, loots and much more...
The dungeon is now far closer to what it was on retail. A special thanks to our newest team member Tobschinski who did most of the rework for these two dungeons.

Many gossips, quests, items, NPCs, events and dungeons were also fixed as usual.Please, read the CHANGELOG.txt file for a full list of changes in version 1.4 since version 1.3.