[0709] Released Classic DB version 1.4.1
Classic DB release version 1.4.1 'Toward the Fiery Heart'

This release is compatible with revision z2525 of CMaNGOS-classic and comes
with revision z2703 of SD2-classic and with last updates of
ACID-classic.

Thanks to all the contributors for their reports, fixes and tests.
As always, very special thanks to the ScriptDev2, ACID, TBC-DB and the
UDB teams as this release would not have been possible without their
contributions.

The main purpose of this intermediate release is to update the ACID version of the
database.

Many gossips, quests, items, NPCs, events, dungeons and raids were also
fixed. Please, read the CHANGELOG.txt file for a full list of changes in
version 1.4.1 since version 1.4.0.