[1229] 2nd Cleanup Update
This is hopefully the final update for this PR which aims to cleanup the
database by moving creature_addon entries to creature_template_addon
entries for
- Midsummer Celebrants
- Firebreathers
- Fireeaters
- Flamekeepers
- Loremasters
- Talespinners
Also removed gameobjects and npcs which were only present in later
versions on the festival including alliance, horde and ahune bonfires,
midsummer bonfire and bonfire despawner npcs.
Finally the alliance stranglethorn midsummer station was removed which
was part of a later version of the event.

Signed-off-by: Cala <calaftp@free.fr>