[1405] Removed wrong spawns of NPC 5890 in Mulgore
* Removed wrong spawns of NPC 5890 (Redrock Earth Spirit) in Mulgore
There should be three small spirits in Durotar and Mulgore for the
shaman quest "Call of the Earth".
There are only two each that are working as intended, plus three
additional bugged ones in Mulgore.
The bugged ones are visible to every class/race without having used the
quest item.
This closes #870

Signed-off-by: Cala <calaftp@free.fr>