[1472] Added support and scripts for Wickerman Festival
* This is the event that occurs at 8pm every evening. Preceding this
are reminders at 30, 15 and 5 minutes before the event starts by
Darkcaller Yanka.
Upon starting, Sylvanas is summoned as the Banshee Queen and delivers
her speech after which, wickerman embers are summoned around the
wickerman (which give a buff) and Darkcaller Yanka starts dancing. This
lasts the night when the dancing stops and the embers despawn.

Signed-off-by: Cala <calaftp@free.fr>