[1518] Applied CMaNGOS core update z2689
Give world database control over spell charges on items. A negative
value signifies that the item should be deleted once all charges are
consumed. This rendered the ITEM_EXTRA_NON_CONSUMABLE flag obsolete,
therefore it was also removed.

WARNING: Future readers may feel that this is bad design.  You're right.
It is.  Unfortunately, it is the behavior the client expects.

Signed-off-by: Cala <calaftp@free.fr>