[1530] Applied CMaNGOS core update z2690
Update game_event table structure to use better time format

This brings two fixes:
1- change the start_time and end_time columns to use DATETIME format
instead of timestamp. Timestamp is stored as an absolute time and is
subject to time zone changes, shifting the effective  start/end time of
an event. Thanks to @Neotmiren for pointing and providing a fix.
2- Change the default value for start_time and end_time column to UNIX
origin of time. The previous zero value was not compliant with SQL
strict mode, thanks @kartoffelsup for reporting. Moreover, it has no
meaning time-wise and could lead to some bad code.