Fix for incorrectly scaled Imp Minion guardians
Imp Minion Guardians were not correctly scaled and ended up with 3 times
as much health as their handler.
This was due to not using the CLS system for calculating stats, but
instead falling back to the old system with linear scaling.
This commit changes the "ArmorMultiplier" from -1 to 0 for Imp Minions,
causing the CLS system to be used instead.