[2640] Finalise CLS activation
* Activate Class Level Stats System for all creatures that were not using it yet.
Thanks @ratkosrb for his help in checking remaining health and mana multipliers against retail ones.

This means that all creatures and pets now have their HP and mana calculated on the fly by the core.
Damage and armor can still be off for some of them and further work is needed and intended on that matter.
Specific cases will be handled one by one.