[3032][TDB]  Implement Engineering specialization re-learning with Book "Soothsaying for Dummies"
To switch spec In Classic and TBC, you are required to drop your Engineering skill entirely, and level it again from 1 to 225. Then the book should allow you to pick a new specialization free of charge.

In early wotlk it was changed so that you can "unlearn" your current specialization with the book for a 150g fee in addition to re-learning. (WotLK-DB should uncomment and implement the "unlearn" gossip options here)