Complete Overhaul of Mining Node Data
I noticed that the spawning on mining nodes in Durotar and The Barrens seemed wonky, so I have spent the last several months working on a total overhaul of the placement of all of the gameobjects and their pooling data for Kalimdor and the Eastern Kingdoms.

I've split this into 3 parts:
4494) Remove all of the old data, including gameobject entries, pooling data, and the spawn_group table entries for random nodes throughout Kal/EK
4995) Insert gameobject and pooling data for EK (map 0)
4996) Insert gameobject and pooling data for Kal (map 1)

The spawning of mining nodes should be much more blizzlike now I believe. The Barrens before my changes only had about 20-30 mining nodes despite it being a huge zone, and the pooling of the nodes that were there was very clearly wrong.

I got the locations of each node from WoWhead and the pooling and spawn chances (% chance a rare node like Silver or Gold will be chosen) were mostly guesswork but I do not know how or where I'd find a reliable source on those numbers.

This will insert/change about 22,000 rows of data in the database, so hopefully this rework is solid and welcome. I built a tool to help visualize the placement and pooling of the nodes in each zone, without which this task would've probably been impossible. I'm guessing that's why nobody had done something like this previously. Here's the repository I created for the tool, I'm kinda dumb with Git so hopefully whatever you need is there, I developed the tool using Eclipse by the way.

https://github.com/Sneakyabc/CmangosPoolViewer

Thanks.